refactor: move AppInfo to ors-engine
