+++ ors-api/src/main/java/org/heigit/ors/api/errors/CommonResponseEntityExceptionHandler.java
@@ -17,7 +17,7 @@ package org.heigit.ors.api.errors;
 
 import com.fasterxml.jackson.databind.exc.ValueInstantiationException;
 import org.apache.log4j.Logger;
-import org.heigit.ors.api.util.AppInfo;
+import org.heigit.ors.util.AppInfo;
 import org.heigit.ors.exceptions.ParameterValueException;
 import org.heigit.ors.exceptions.StatusCodeException;
 import org.heigit.ors.exceptions.UnknownParameterException;
