+++ ors-api/src/main/java/org/heigit/ors/api/responses/isochrones/IsochronesResponseInfo.java
@@ -24,7 +24,7 @@ import org.heigit.ors.api.config.EndpointsProperties;
 import org.heigit.ors.api.config.SystemMessageProperties;
 import org.heigit.ors.api.requests.isochrones.IsochronesRequest;
 import org.heigit.ors.api.responses.common.engineinfo.EngineInfo;
-import org.heigit.ors.api.util.AppInfo;
+import org.heigit.ors.util.AppInfo;
 import org.heigit.ors.api.util.SystemMessage;
 
 @Schema(name = "IsochronesResponseInfo", description = "Information about the request")
