+++ ors-api/src/main/java/org/heigit/ors/api/responses/routing/gpx/GPXExtensions.java
@@ -18,7 +18,7 @@ package org.heigit.ors.api.responses.routing.gpx;
 import com.graphhopper.util.Helper;
 import jakarta.xml.bind.annotation.XmlElement;
 import org.heigit.ors.api.requests.routing.RouteRequest;
-import org.heigit.ors.api.util.AppInfo;
+import org.heigit.ors.util.AppInfo;
 import org.heigit.ors.api.APIEnums;
 
 public class GPXExtensions {
