+++ ors-api/src/main/java/org/heigit/ors/api/services/GraphService.java
@@ -1,7 +1,7 @@
 package org.heigit.ors.api.services;
 
 import org.apache.log4j.Logger;
-import org.heigit.ors.api.util.AppInfo;
+import org.heigit.ors.util.AppInfo;
 import org.heigit.ors.config.EngineProperties;
 import org.heigit.ors.routing.RoutingProfile;
 import org.heigit.ors.routing.RoutingProfileManager;
